+++ include/dnn/dnn_math.h
@@ -22,6 +22,6 @@
 #include <stdio.h>
 #define DNN_EULER 2.718281828459045
 
-float sigmoid(float x);
+double sigmoid(double x);
 
 #endif
